% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assignMedoids}
\alias{assignMedoids}
\title{assignMedoids}
\usage{
assignMedoids(curves, medoids)
}
\arguments{
\item{curves}{(Chunk) of series whose medoids indices must be found}

\item{medoids}{Matrix of medoids (in columns)}
}
\value{
The vector of integer assignments
}
\description{
Find the closest medoid for each curve in input
}
