% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanBin}
\alias{cleanBin}
\title{cleanBin}
\usage{
cleanBin()
}
\description{
Remove binary files to re-generate them at next run of \code{claws()}.
To be run in the folder where computations occurred (or no effect).
}
