% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSynchrones.R
\name{computeSynchrones}
\alias{computeSynchrones}
\title{computeSynchrones}
\usage{
computeSynchrones(medoids, getSeries, nb_curves, nb_series_per_chunk,
  ncores = 3, verbose = FALSE)
}
\arguments{
\item{medoids}{matrix of K medoids curves in columns}

\item{getSeries}{Function to retrieve series (argument: 'inds', integer vector),
as columns of a matrix}

\item{nb_curves}{How many series? (this is known, at this stage)}

\item{nb_series_per_chunk}{Number of series to retrieve in one batch}

\item{ncores}{Number of cores for parallel runs}

\item{verbose}{FALSE: nothing printed; TRUE: some execution traces}
}
\value{
A matrix of K synchrones in columns (same length as the series)
}
\description{
Compute the synchrones curves (sums of clusters elements) from a matrix of medoids,
using euclidian distance.
}
