% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeWerDists.R
\name{computeWerDists}
\alias{computeWerDists}
\title{computeWerDists}
\usage{
computeWerDists(indices, getSeries, nb_series_per_chunk, smooth_lvl = 3,
  nvoice = 4, nbytes = 4, endian = .Platform$endian, ncores = 3,
  verbose = FALSE)
}
\arguments{
\item{indices}{Indices of the series to consider}

\item{getSeries}{Function to retrieve series (argument: 'inds', integer vector),
as columns of a matrix}

\item{nb_series_per_chunk}{Number of series to retrieve in one batch}

\item{smooth_lvl}{Smoothing level: odd integer, 1 == no smoothing.}

\item{nvoice}{Number of voices within each octave for CWT computations}

\item{nbytes}{4 or 8 bytes to (de)serialize a floating-point number}

\item{endian}{Endianness for (de)serialization: "little" or "big"}

\item{ncores}{Number of cores for parallel runs}

\item{verbose}{FALSE: nothing printed; TRUE: some execution traces}
}
\value{
A distances matrix of size K x K where K == length(indices)
}
\description{
Compute the WER distances between the series at specified indices, which are
obtaind by \code{getSeries(indices)}
}
