% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{curvesToContribs}
\alias{curvesToContribs}
\title{curvesToContribs}
\usage{
curvesToContribs(curves, wav_filt, contrib_type)
}
\arguments{
\item{curves}{[big.]matrix of series (in columns), of size L x n}

\item{wav_filt}{Wavelet transform filter, as a vector c(Family,FilterNumber)}

\item{contrib_type}{Type of contribution: "relative", "logit" or "absolute" (any
prefix)}
}
\value{
A matrix of size log(L) x n containing contributions in columns
}
\description{
Compute the discrete wavelet coefficients for each series, and aggregate them in
energy contribution across scales as described in https://arxiv.org/abs/1101.4744v2
}
