% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterMA}
\alias{filterMA}
\title{filterMA}
\usage{
filterMA(M_, w_)
}
\arguments{
\item{M_}{A real matrix of size LxD}

\item{w_}{The (odd) number of values to average}
}
\value{
The filtered matrix (in columns), of same size as the input
}
\description{
Filter [time-]series by replacing all values by the moving average of values
centered around current one. Border values are averaged with available data.
}
