% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateCurve.R
\name{getDetRandCurve}
\alias{getDetRandCurve}
\title{getDetRandCurve}
\usage{
getDetRandCurve(indices)
}
\arguments{
\item{i}{Some index}
}
\description{
TODO: improve this (using known clusters centers...)
}
\examples{
\dontrun{
initCentroids(N=250000, D=15000, K=15)
res_ascii <- claws(getDetRandCurve, K1=50, K2=15, nb_series_per_chunk=500,
  nb_items_clust=100, random=FALSE, verbose=TRUE, ncores_clust=3)
}
}
