% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateCurve.R
\name{initClustersParams}
\alias{initClustersParams}
\title{initCentroids}
\usage{
initClustersParams(N, D, K)
}
\arguments{
\item{N}{Number of curves}

\item{D}{Number of sample points}

\item{K}{Number of clusters}
}
\description{
Must be called before getDetRandCurve()
}
